+++ tests/models/readout_electronics/test_simple_adc.py
@@ -28,8 +28,6 @@ def ccd_3x3() -> CCD:
         environment=Environment(),
         characteristics=CCDCharacteristics(),
     )
-    detector.characteristics.adc_bit_resolution = 16
-    detector.characteristics.adc_voltage_range = (0.0, 5.0)
     return detector
 
 
@@ -40,35 +38,19 @@ def test_simple_adc(ccd_3x3: CCD, data_type: str):
 
 
 @pytest.mark.parametrize(
-    "voltage_range, bit_resolution, data_type, exp_exc, exp_error",
+    "data_type, exp_exc, exp_error",
     [
-        ([0.0, 5.0], 16, "float", TypeError, "Expecting a signed/unsigned integer"),
-        ([0.0, 5.0], 16, "float32", TypeError, "Expecting a signed/unsigned integer"),
-        ([0.0, 5.0], 16, "float64", TypeError, "Expecting a signed/unsigned integer"),
-        ([0.0, 5.0], 16, "uint8", TypeError, "Expected type of Image array is uint64"),
-        ([0.0, 5.0], 16, "int32", TypeError, "Expected type of Image array is uint64"),
-        ([0.0, 5.0], 16, "int64", TypeError, "Expected type of Image array is uint64"),
-        ([0.0, 5.0], 16, "int", TypeError, "Expected type of Image array is uint64"),
-        (
-            [0.0, 5.0, 6.0],
-            16,
-            "int",
-            ValueError,
-            "Voltage range must have length of 2.",
-        ),
-        (
-            [0.0, 5.0],
-            100,
-            "int",
-            ValueError,
-            "'adc_bit_resolution' must be between 4 and 64.",
-        ),
+        ("float", TypeError, "Expecting a signed/unsigned integer"),
+        ("float32", TypeError, "Expecting a signed/unsigned integer"),
+        ("float64", TypeError, "Expecting a signed/unsigned integer"),
+        ("uint8", TypeError, "Expected type of Image array is uint64"),
+        ("int32", TypeError, "Expected type of Image array is uint64"),
+        ("int64", TypeError, "Expected type of Image array is uint64"),
+        ("int", TypeError, "Expected type of Image array is uint64"),
     ],
 )
 def test_simple_adc_wrong_data_type(
     ccd_3x3: CCD,
-    voltage_range: t.Tuple[float, float],
-    bit_resolution: int,
     data_type: str,
     exp_exc,
     exp_error,
@@ -77,7 +59,5 @@ def test_simple_adc_wrong_data_type(
     with pytest.raises(exp_exc, match=exp_error):
         simple_adc(
             detector=ccd_3x3,
-            voltage_range=voltage_range,
-            bit_resolution=bit_resolution,
             data_type=data_type,
         )